corrected misplacement of cmdline files and latest attempt at diff_dicts method
